Split lib/ast.js into multiple modules in lib/ast/
On the one hand, this makes the code more manageable, on the other
hand, this solves some cyclic import issues. So win win!
