+++ lib/ast/api.js
@@ -11,17 +11,17 @@
 
 const assert = require('assert');
 
-const Ast = require('./ast');
-const Type = require('./type');
-const Builtin = require('./builtin/defs');
-const { prettyprint, prettyprintExample } = require('./prettyprint');
+const Ast = require('.');
+const Type = require('../type');
+const Builtin = require('../builtin/defs');
+const { prettyprint, prettyprintExample } = require('../prettyprint');
 
 const { isUnaryStreamToStreamOp,
         isUnaryTableToTableOp,
         isUnaryStreamToTableOp,
-        isUnaryTableToStreamOp } = require('./utils');
-const { optimizeFilter, optimizeProgram } = require('./optimize');
-let { typeCheckFilter, typeCheckProgram, typeCheckPermissionRule, typeCheckMeta } = require('./typecheck');
+        isUnaryTableToStreamOp } = require('../utils');
+const { optimizeFilter, optimizeProgram } = require('../optimize');
+let { typeCheckFilter, typeCheckProgram, typeCheckPermissionRule, typeCheckMeta } = require('../typecheck');
 
 const ProgramProto = Object.getPrototypeOf(new Ast.Program([], [], [], null));
 const PermissionRuleProto = Object.getPrototypeOf(new Ast.PermissionRule(Ast.BooleanExpression.True, Ast.PermissionFunction.Builtin, Ast.PermissionFunction.Builtin));
