+++ lib/builtin/defs.js
@@ -10,7 +10,7 @@
 "use strict";
 
 const Type = require('../type');
-const { FunctionDef } = require('../class_def_ast');
+const { FunctionDef } = require('../ast/function_def');
 
 // Definitions of ThingTalk operators
