+++ lib/generate.js
@@ -15,7 +15,7 @@ const { isUnaryStreamToStreamOp,
         isUnaryTableToStreamOp } = require('./utils');
 
 // Initialize the AST API
-const { notifyAction } = require('./ast_api');
+const { notifyAction } = require('./ast/api');
 
 module.exports = {
     notifyAction,
