+++ lib/grammar_api.js
@@ -11,7 +11,7 @@
 
 const Grammar = require('./grammar');
 // Initialize the AST API
-require('./ast_api');
+require('./ast/api');
 
 module.exports = {
     parse(code) {
