+++ lib/schema.js
@@ -14,7 +14,8 @@ const assert = require('assert');
 const Type = require('./type');
 const Grammar = require('./grammar_api');
 const { typeCheckClass } = require('./typecheck');
-const Ast = require('./class_def_ast');
+const { ClassDef } = require('./ast/class_def');
+const Ast = require('./ast/function_def');
 
 function delay(timeout) {
     return new Promise((resolve, reject) => {
@@ -74,7 +75,7 @@ module.exports = class SchemaRetriever {
     }
 
     injectManifest(kind, manifest) {
-        const classDef = Ast.ClassDef.fromManifest(kind, manifest);
+        const classDef = ClassDef.fromManifest(kind, manifest);
         this.injectClass(classDef);
     }
     injectClass(classDef) {
